+++ packages/asset-swapper/src/types.ts
@@ -4,7 +4,13 @@ import { TakerRequestQueryParams } from '@0x/quote-server';
 import { SignedOrder } from '@0x/types';
 import { BigNumber } from '@0x/utils';
 
-import { ERC20BridgeSource, GetMarketOrdersOpts, OptimizedMarketOrder } from './utils/market_operation_utils/types';
+import {
+    ERC20BridgeSource,
+    GetMarketOrdersOpts,
+    LiquidityProviderRegistry,
+    OptimizedMarketOrder,
+    TokenAdjacencyGraph,
+} from './utils/market_operation_utils/types';
 import { QuoteReport } from './utils/quote_report_generator';
 
 /**
@@ -310,11 +316,12 @@ export interface SwapQuoterOpts extends OrderPrunerOpts {
     ethereumRpcUrl?: string;
     contractAddresses?: AssetSwapperContractAddresses;
     samplerGasLimit?: number;
-    liquidityProviderRegistryAddress?: string;
     multiBridgeAddress?: string;
     ethGasStationUrl?: string;
     rfqt?: SwapQuoterRfqtOpts;
     samplerOverrides?: SamplerOverrides;
+    tokenAdjacencyGraph?: TokenAdjacencyGraph;
+    liquidityProviderRegistry?: LiquidityProviderRegistry;
 }
 
 /**