+++ packages/asset-swapper/src/utils/market_operation_utils/sampler_operations.ts
@@ -1,17 +1,17 @@
 import { SupportedProvider } from '@0x/dev-utils';
 import { SignedOrder } from '@0x/types';
-import { BigNumber, NULL_ADDRESS } from '@0x/utils';
+import { BigNumber } from '@0x/utils';
 import * as _ from 'lodash';
 
 import { ERC20BridgeSamplerContract } from '../../wrappers';
 
 import { BalancerPoolsCache, computeBalancerBuyQuote, computeBalancerSellQuote } from './balancer_utils';
 import { BancorService } from './bancor_service';
-import { MAINNET_SUSHI_SWAP_ROUTER, MAX_UINT256, NULL_BYTES, ZERO_AMOUNT } from './constants';
+import { LIQUIDITY_PROVIDER_REGISTRY, MAINNET_SUSHI_SWAP_ROUTER, MAX_UINT256, ZERO_AMOUNT } from './constants';
 import { CreamPoolsCache } from './cream_utils';
 import { getCurveInfosForPair, getSnowSwapInfosForPair, getSwerveInfosForPair } from './curve_utils';
 import { getKyberReserveIdsForPair } from './kyber_utils';
-import { getMultiBridgeIntermediateToken } from './multibridge_utils';
+import { getLiquidityProvidersForPair } from './liquidity_provider_utils';
 import { getIntermediateTokens } from './multihop_utils';
 import { SamplerContractOperation } from './sampler_contract_operation';
 import { getShellsForPair } from './shell_utils';
@@ -28,8 +28,8 @@ import {
     HopInfo,
     KyberFillData,
     LiquidityProviderFillData,
+    LiquidityProviderRegistry,
     MooniswapFillData,
-    MultiBridgeFillData,
     MultiHopFillData,
     ShellFillData,
     SnowSwapFillData,
@@ -47,7 +47,6 @@ import {
  */
 export const TWO_HOP_SOURCE_FILTERS = SourceFilters.all().exclude([
     ERC20BridgeSource.MultiHop,
-    ERC20BridgeSource.MultiBridge,
     ERC20BridgeSource.Native,
 ]);
 /**
@@ -80,6 +79,8 @@ export class SamplerOperations {
         public readonly balancerPoolsCache: BalancerPoolsCache = new BalancerPoolsCache(),
         public readonly creamPoolsCache: CreamPoolsCache = new CreamPoolsCache(),
         protected readonly getBancorServiceFn?: () => BancorService, // for dependency injection in tests
+        protected readonly tokenAdjacencyGraph: TokenAdjacencyGraph = { default: [] },
+        public readonly liquidityProviderRegistry: LiquidityProviderRegistry = LIQUIDITY_PROVIDER_REGISTRY,
     ) {}
 
     public async getBancorServiceAsync(): Promise<BancorService> {
@@ -221,64 +222,32 @@ export class SamplerOperations {
     }
 
     public getLiquidityProviderSellQuotes(
-        registryAddress: string,
+        providerAddress: string,
         makerToken: string,
         takerToken: string,
         takerFillAmounts: BigNumber[],
     ): SourceQuoteOperation<LiquidityProviderFillData> {
         return new SamplerContractOperation({
             source: ERC20BridgeSource.LiquidityProvider,
-            fillData: {} as LiquidityProviderFillData, // tslint:disable-line:no-object-literal-type-assertion
+            fillData: { poolAddress: providerAddress },
             contract: this._samplerContract,
-            function: this._samplerContract.sampleSellsFromLiquidityProviderRegistry,
-            params: [registryAddress, takerToken, makerToken, takerFillAmounts],
-            callback: (callResults: string, fillData: LiquidityProviderFillData): BigNumber[] => {
-                const [samples, poolAddress] = this._samplerContract.getABIDecodedReturnData<[BigNumber[], string]>(
-                    'sampleSellsFromLiquidityProviderRegistry',
-                    callResults,
-                );
-                fillData.poolAddress = poolAddress;
-                return samples;
-            },
+            function: this._samplerContract.sampleSellsFromLiquidityProvider,
+            params: [providerAddress, takerToken, makerToken, takerFillAmounts],
         });
     }
 
     public getLiquidityProviderBuyQuotes(
-        registryAddress: string,
+        providerAddress: string,
         makerToken: string,
         takerToken: string,
         makerFillAmounts: BigNumber[],
     ): SourceQuoteOperation<LiquidityProviderFillData> {
         return new SamplerContractOperation({
             source: ERC20BridgeSource.LiquidityProvider,
-            fillData: {} as LiquidityProviderFillData, // tslint:disable-line:no-object-literal-type-assertion
+            fillData: { poolAddress: providerAddress },
             contract: this._samplerContract,
-            function: this._samplerContract.sampleBuysFromLiquidityProviderRegistry,
-            params: [registryAddress, takerToken, makerToken, makerFillAmounts],
-            callback: (callResults: string, fillData: LiquidityProviderFillData): BigNumber[] => {
-                const [samples, poolAddress] = this._samplerContract.getABIDecodedReturnData<[BigNumber[], string]>(
-                    'sampleBuysFromLiquidityProviderRegistry',
-                    callResults,
-                );
-                fillData.poolAddress = poolAddress;
-                return samples;
-            },
-        });
-    }
-
-    public getMultiBridgeSellQuotes(
-        multiBridgeAddress: string,
-        makerToken: string,
-        intermediateToken: string,
-        takerToken: string,
-        takerFillAmounts: BigNumber[],
-    ): SourceQuoteOperation<MultiBridgeFillData> {
-        return new SamplerContractOperation({
-            source: ERC20BridgeSource.MultiBridge,
-            fillData: { poolAddress: multiBridgeAddress },
-            contract: this._samplerContract,
-            function: this._samplerContract.sampleSellsFromMultiBridge,
-            params: [multiBridgeAddress, takerToken, intermediateToken, makerToken, takerFillAmounts],
+            function: this._samplerContract.sampleBuysFromLiquidityProvider,
+            params: [providerAddress, takerToken, makerToken, makerFillAmounts],
         });
     }
 
@@ -669,34 +638,15 @@ export class SamplerOperations {
         makerToken: string,
         takerToken: string,
         sellAmount: BigNumber,
-        wethAddress: string,
-        tokenAdjacencyGraph: TokenAdjacencyGraph,
-        liquidityProviderRegistryAddress?: string,
     ): BatchedOperation<Array<DexSample<MultiHopFillData>>> {
         const _sources = TWO_HOP_SOURCE_FILTERS.getAllowed(sources);
         if (_sources.length === 0) {
             return SamplerOperations.constant([]);
         }
-        const intermediateTokens = getIntermediateTokens(makerToken, takerToken, tokenAdjacencyGraph, wethAddress);
+        const intermediateTokens = getIntermediateTokens(makerToken, takerToken, this.tokenAdjacencyGraph);
         const subOps = intermediateTokens.map(intermediateToken => {
-            const firstHopOps = this._getSellQuoteOperations(
-                _sources,
-                intermediateToken,
-                takerToken,
-                [ZERO_AMOUNT],
-                wethAddress,
-                tokenAdjacencyGraph,
-                liquidityProviderRegistryAddress,
-            );
-            const secondHopOps = this._getSellQuoteOperations(
-                _sources,
-                makerToken,
-                intermediateToken,
-                [ZERO_AMOUNT],
-                wethAddress,
-                tokenAdjacencyGraph,
-                liquidityProviderRegistryAddress,
-            );
+            const firstHopOps = this._getSellQuoteOperations(_sources, intermediateToken, takerToken, [ZERO_AMOUNT]);
+            const secondHopOps = this._getSellQuoteOperations(_sources, makerToken, intermediateToken, [ZERO_AMOUNT]);
             return new SamplerContractOperation({
                 contract: this._samplerContract,
                 source: ERC20BridgeSource.MultiHop,
@@ -747,34 +697,19 @@ export class SamplerOperations {
         makerToken: string,
         takerToken: string,
         buyAmount: BigNumber,
-        wethAddress: string,
-        tokenAdjacencyGraph: TokenAdjacencyGraph,
-        liquidityProviderRegistryAddress?: string,
     ): BatchedOperation<Array<DexSample<MultiHopFillData>>> {
         const _sources = TWO_HOP_SOURCE_FILTERS.getAllowed(sources);
         if (_sources.length === 0) {
             return SamplerOperations.constant([]);
         }
-        const intermediateTokens = getIntermediateTokens(makerToken, takerToken, tokenAdjacencyGraph, wethAddress);
+        const intermediateTokens = getIntermediateTokens(makerToken, takerToken, this.tokenAdjacencyGraph);
         const subOps = intermediateTokens.map(intermediateToken => {
-            const firstHopOps = this._getBuyQuoteOperations(
-                _sources,
-                intermediateToken,
-                takerToken,
-                [new BigNumber(0)],
-                wethAddress,
-                tokenAdjacencyGraph, // TODO is this a bad idea?
-                liquidityProviderRegistryAddress,
-            );
-            const secondHopOps = this._getBuyQuoteOperations(
-                _sources,
-                makerToken,
-                intermediateToken,
-                [new BigNumber(0)],
-                wethAddress,
-                tokenAdjacencyGraph,
-                liquidityProviderRegistryAddress,
-            );
+            const firstHopOps = this._getBuyQuoteOperations(_sources, intermediateToken, takerToken, [
+                new BigNumber(0),
+            ]);
+            const secondHopOps = this._getBuyQuoteOperations(_sources, makerToken, intermediateToken, [
+                new BigNumber(0),
+            ]);
             return new SamplerContractOperation({
                 contract: this._samplerContract,
                 source: ERC20BridgeSource.MultiHop,
@@ -853,17 +788,10 @@ export class SamplerOperations {
     ): SourceQuoteOperation<ShellFillData> {
         return new SamplerContractOperation({
             source: ERC20BridgeSource.Shell,
+            fillData: { poolAddress },
             contract: this._samplerContract,
             function: this._samplerContract.sampleSellsFromShell,
             params: [poolAddress, takerToken, makerToken, takerFillAmounts],
-            callback: (callResults: string, fillData: ShellFillData): BigNumber[] => {
-                const samples = this._samplerContract.getABIDecodedReturnData<BigNumber[]>(
-                    'sampleSellsFromShell',
-                    callResults,
-                );
-                fillData.poolAddress = poolAddress;
-                return samples;
-            },
         });
     }
 
@@ -875,17 +803,10 @@ export class SamplerOperations {
     ): SourceQuoteOperation {
         return new SamplerContractOperation({
             source: ERC20BridgeSource.Shell,
+            fillData: { poolAddress },
             contract: this._samplerContract,
             function: this._samplerContract.sampleBuysFromShell,
             params: [poolAddress, takerToken, makerToken, makerFillAmounts],
-            callback: (callResults: string, fillData: ShellFillData): BigNumber[] => {
-                const samples = this._samplerContract.getABIDecodedReturnData<BigNumber[]>(
-                    'sampleBuysFromShell',
-                    callResults,
-                );
-                fillData.poolAddress = poolAddress;
-                return samples;
-            },
         });
     }
 
@@ -936,52 +857,36 @@ export class SamplerOperations {
         makerToken: string,
         takerToken: string,
         takerFillAmount: BigNumber,
-        liquidityProviderRegistryAddress?: string,
-        multiBridgeAddress?: string,
     ): BatchedOperation<BigNumber> {
         if (makerToken.toLowerCase() === takerToken.toLowerCase()) {
             return SamplerOperations.constant(new BigNumber(1));
         }
-        const getSellQuotes = this.getSellQuotes(
-            sources,
-            makerToken,
-            takerToken,
-            [takerFillAmount],
-            NULL_ADDRESS, // weth address
-            {}, // token adjacency
-            liquidityProviderRegistryAddress,
-            multiBridgeAddress,
-        );
+        const subOps = this._getSellQuoteOperations(sources, makerToken, takerToken, [takerFillAmount], {
+            default: [],
+        });
         return {
             encodeCall: () => {
-                const encodedCall = getSellQuotes.encodeCall();
-                // All soures were excluded
-                if (encodedCall === NULL_BYTES) {
-                    return NULL_BYTES;
-                }
-                return this._samplerContract.batchCall([encodedCall]).getABIEncodedTransactionData();
+                const subCalls = subOps.map(op => op.encodeCall());
+                return this._samplerContract.batchCall(subCalls).getABIEncodedTransactionData();
             },
             handleCallResults: callResults => {
-                if (callResults === NULL_BYTES) {
-                    return ZERO_AMOUNT;
-                }
                 const rawSubCallResults = this._samplerContract.getABIDecodedReturnData<string[]>(
                     'batchCall',
                     callResults,
                 );
-                const samples = getSellQuotes.handleCallResults(rawSubCallResults[0]);
+                const samples = subOps.map((op, i) => op.handleCallResults(rawSubCallResults[i]));
                 if (samples.length === 0) {
                     return ZERO_AMOUNT;
                 }
                 const flatSortedSamples = samples
                     .reduce((acc, v) => acc.concat(...v))
-                    .filter(v => !v.output.isZero())
-                    .sort((a, b) => a.output.comparedTo(b.output));
+                    .filter(v => !v.isZero())
+                    .sort((a, b) => a.comparedTo(b));
                 if (flatSortedSamples.length === 0) {
                     return ZERO_AMOUNT;
                 }
                 const medianSample = flatSortedSamples[Math.floor(flatSortedSamples.length / 2)];
-                return medianSample.output.div(medianSample.input);
+                return medianSample.div(takerFillAmount);
             },
         };
     }
@@ -991,21 +896,8 @@ export class SamplerOperations {
         makerToken: string,
         takerToken: string,
         takerFillAmounts: BigNumber[],
-        wethAddress: string,
-        tokenAdjacencyGraph: TokenAdjacencyGraph,
-        liquidityProviderRegistryAddress?: string,
-        multiBridgeAddress?: string,
     ): BatchedOperation<DexSample[][]> {
-        const subOps = this._getSellQuoteOperations(
-            sources,
-            makerToken,
-            takerToken,
-            takerFillAmounts,
-            wethAddress,
-            tokenAdjacencyGraph,
-            liquidityProviderRegistryAddress,
-            multiBridgeAddress,
-        );
+        const subOps = this._getSellQuoteOperations(sources, makerToken, takerToken, takerFillAmounts);
         return {
             encodeCall: () => {
                 const subCalls = subOps.map(op => op.encodeCall());
@@ -1034,19 +926,8 @@ export class SamplerOperations {
         makerToken: string,
         takerToken: string,
         makerFillAmounts: BigNumber[],
-        wethAddress: string,
-        tokenAdjacencyGraph: TokenAdjacencyGraph,
-        liquidityProviderRegistryAddress?: string,
     ): BatchedOperation<DexSample[][]> {
-        const subOps = this._getBuyQuoteOperations(
-            sources,
-            makerToken,
-            takerToken,
-            makerFillAmounts,
-            wethAddress,
-            tokenAdjacencyGraph,
-            liquidityProviderRegistryAddress,
-        );
+        const subOps = this._getBuyQuoteOperations(sources, makerToken, takerToken, makerFillAmounts);
         return {
             encodeCall: () => {
                 const subCalls = subOps.map(op => op.encodeCall());
@@ -1075,21 +956,12 @@ export class SamplerOperations {
         makerToken: string,
         takerToken: string,
         takerFillAmounts: BigNumber[],
-        wethAddress: string,
-        tokenAdjacencyGraph: TokenAdjacencyGraph,
-        liquidityProviderRegistryAddress?: string,
-        multiBridgeAddress?: string,
+        tokenAdjacencyGraph: TokenAdjacencyGraph = this.tokenAdjacencyGraph,
     ): SourceQuoteOperation[] {
-        const _sources = BATCH_SOURCE_FILTERS.exclude(
-            liquidityProviderRegistryAddress ? [] : [ERC20BridgeSource.LiquidityProvider],
-        )
-            .exclude(multiBridgeAddress || multiBridgeAddress === NULL_ADDRESS ? [] : [ERC20BridgeSource.MultiBridge])
-            .getAllowed(sources);
-
         // Find the adjacent tokens in the provided tooken adjacency graph,
         // e.g if this is DAI->USDC we may check for DAI->WETH->USDC
-        const intermediateTokens = getIntermediateTokens(makerToken, takerToken, tokenAdjacencyGraph, wethAddress);
-
+        const intermediateTokens = getIntermediateTokens(makerToken, takerToken, tokenAdjacencyGraph);
+        const _sources = BATCH_SOURCE_FILTERS.getAllowed(sources);
         return _.flatten(
             _sources.map(
                 (source): SourceQuoteOperation | SourceQuoteOperation[] => {
@@ -1144,30 +1016,12 @@ export class SamplerOperations {
                                 ),
                             );
                         case ERC20BridgeSource.LiquidityProvider:
-                            if (liquidityProviderRegistryAddress === undefined) {
-                                throw new Error(
-                                    'Cannot sample liquidity from a LiquidityProvider liquidity pool, if a registry is not provided.',
-                                );
-                            }
-                            return this.getLiquidityProviderSellQuotes(
-                                liquidityProviderRegistryAddress,
-                                makerToken,
+                            return getLiquidityProvidersForPair(
+                                this.liquidityProviderRegistry,
                                 takerToken,
-                                takerFillAmounts,
-                            );
-                        case ERC20BridgeSource.MultiBridge:
-                            if (multiBridgeAddress === undefined) {
-                                throw new Error(
-                                    'Cannot sample liquidity from MultiBridge if an address is not provided.',
-                                );
-                            }
-                            const intermediateToken = getMultiBridgeIntermediateToken(takerToken, makerToken);
-                            return this.getMultiBridgeSellQuotes(
-                                multiBridgeAddress,
                                 makerToken,
-                                intermediateToken,
-                                takerToken,
-                                takerFillAmounts,
+                            ).map(pool =>
+                                this.getLiquidityProviderSellQuotes(pool, makerToken, takerToken, takerFillAmounts),
                             );
                         case ERC20BridgeSource.MStable:
                             return this.getMStableSellQuotes(makerToken, takerToken, takerFillAmounts);
@@ -1216,18 +1070,11 @@ export class SamplerOperations {
         makerToken: string,
         takerToken: string,
         makerFillAmounts: BigNumber[],
-        wethAddress: string,
-        tokenAdjacencyGraph: TokenAdjacencyGraph,
-        liquidityProviderRegistryAddress?: string,
     ): SourceQuoteOperation[] {
-        const _sources = BATCH_SOURCE_FILTERS.exclude(
-            liquidityProviderRegistryAddress ? [] : [ERC20BridgeSource.LiquidityProvider],
-        ).getAllowed(sources);
-
         // Find the adjacent tokens in the provided tooken adjacency graph,
         // e.g if this is DAI->USDC we may check for DAI->WETH->USDC
-        const intermediateTokens = getIntermediateTokens(makerToken, takerToken, tokenAdjacencyGraph, wethAddress);
-
+        const intermediateTokens = getIntermediateTokens(makerToken, takerToken, this.tokenAdjacencyGraph);
+        const _sources = BATCH_SOURCE_FILTERS.getAllowed(sources);
         return _.flatten(
             _sources.map(
                 (source): SourceQuoteOperation | SourceQuoteOperation[] => {
@@ -1282,16 +1129,12 @@ export class SamplerOperations {
                                 ),
                             );
                         case ERC20BridgeSource.LiquidityProvider:
-                            if (liquidityProviderRegistryAddress === undefined) {
-                                throw new Error(
-                                    'Cannot sample liquidity from a LiquidityProvider liquidity pool, if a registry is not provided.',
-                                );
-                            }
-                            return this.getLiquidityProviderBuyQuotes(
-                                liquidityProviderRegistryAddress,
-                                makerToken,
+                            return getLiquidityProvidersForPair(
+                                this.liquidityProviderRegistry,
                                 takerToken,
-                                makerFillAmounts,
+                                makerToken,
+                            ).map(pool =>
+                                this.getLiquidityProviderBuyQuotes(pool, makerToken, takerToken, makerFillAmounts),
                             );
                         case ERC20BridgeSource.MStable:
                             return this.getMStableBuyQuotes(makerToken, takerToken, makerFillAmounts);