+++ packages/order-utils/CHANGELOG.md
@@ -5,6 +5,10 @@ Edit the package's CHANGELOG.json file only.
 
 CHANGELOG
 
+## v10.4.6 - _November 13, 2020_
+
+    * Dependencies updated
+
 ## v10.4.5 - _November 3, 2020_
 
     * Dependencies updated